lab 8: added unit tests for multiplication, changed matrix element type to int64

+++ lab_08_04/inc/matrix.h
@@ -10,10 +10,10 @@ typedef struct
     uint32_t rows;
     uint32_t cols;
 
-    int **element;
+    int64_t **element;
 } matrix_t;
 
-exit_t matrix_initialize(int ***matrix, uint32_t rows, uint32_t cols);
+exit_t matrix_initialize(int64_t ***matrix, uint32_t rows, uint32_t cols);
 exit_t matrix_to_square(matrix_t *matrix);
 exit_t matrix_to_same_size(matrix_t *first, matrix_t *second);
 exit_t matrix_pow(matrix_t *matrix, uint32_t power);
